BUG fixed
