--- bin/.gitignore
@@ -1,2 +1,0 @@
-/project0.class
-/Project0.class
